refactor: AnswerEntity test refactoring
